proj_med_img/requirements.txt
annotated-doc==0.0.4
asttokens==3.0.1
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.3.2
colorama==0.4.6
comm==0.2.3
contourpy==1.3.3
cycler==0.12.1
debugpy==1.8.20
decorator==5.2.1
executing==2.2.1
filelock==3.25.2
filetype==1.2.0
flatbuffers==25.12.19
fonttools==4.62.1
fsspec==2026.2.0
idna==3.7
ipykernel==7.2.0
ipython==9.12.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.3
jupyter_client==8.8.0
jupyter_core==5.9.1
kiwisolver==1.5.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
matplotlib-inline==0.2.1
mdurl==0.1.2
ml_dtypes==0.5.4
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.6.1
numpy==2.4.4
onnx==1.21.0
onnxruntime==1.24.4
onnxslim==0.1.90
opencv-python==4.13.0.92
opencv-python-headless==4.10.0.84
packaging==26.0
pandas==3.0.2
parso==0.8.6
pexpect==4.9.0
pi_heif==1.3.0
pillow==12.2.0
pillow-avif-plugin==1.5.5
platformdirs==4.9.4
polars==1.39.3
polars-runtime-32==1.39.3
prompt_toolkit==3.0.52
protobuf==7.34.1
psutil==7.2.2
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.20.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
PyYAML==6.0.3
pyzmq==27.1.0
requests==2.33.1
requests-toolbelt==1.0.0
rich==14.3.3
roboflow==1.3.1
scikit-learn==1.8.0
scipy==1.17.1
seaborn==0.13.2
setuptools==70.2.0
shellingham==1.5.4
six==1.17.0
stack-data==0.6.3
sympy==1.14.0
threadpoolctl==3.6.0
torch==2.11.0+cpu
torchaudio==2.11.0+cpu
torchvision==0.26.0+cpu
tornado==6.5.5
tqdm==4.67.3
traitlets==5.14.3
typer==0.24.1
typing_extensions==4.15.0
-e git+https://github.com/ultralytics/ultralytics.git@8d4e6e841c89f6598b322695cb2bc816eeba8b93#egg=ultralytics
ultralytics-thop==2.0.18
urllib3==2.6.3
wcwidth==0.6.0
